Try with a multiplier and frequency map
This attempt loops over all bus IDs, multiplying them individually with
a multiplier that increases by 1 over each iteration. It stores the
result in a frequency map that keeps track of how many times the given
timestamp has been seen. If the same (offset-adjusted) timestamp has
been seen by all bus lanes, i.e. the freq[timestamp] equals number of
lanes, then that is the correct result.

This again works for the test case, but turns out to be too slow for the
real input.