Implement fast solution for 2020 day 13 part 2
This runs in a few milliseconds.

Each ID (5, 2, 3, 7 in the example below) seems to align with its
previous ID at an interval equal to the result of multiplying the two.

The solution leverages this property by maintaining a step counter that
multiplies by the last ID for each successive bus lane whenever an
alignment has been found. For each iteration, aligned or not, the
timestamp is incremented by the current step value.

The value of the timestamp after iterating over all bus lanes is the
answer. The answer for the example below is 25.

t  | 5  2  3  7
---------------
00 | D  D  D  D
01 | .  .  .  .
02 | .  D  .  .  <- (2, 3) aligned for the first time
03 | .  .  D  .
04 | .  D  .  .
05 | D  .  .  .  <- (5, 2) aligned for the first time
06 | .  D  D  .  <- (3, 7) aligned for the first time
07 | .  .  .  D
08 | .  D  .  .  <- (2, 3) aligned; 2*3 = 6 steps since previous
09 | .  .  D  .
10 | D  D  .  .
11 | .  .  .  .
12 | .  D  D  .
13 | .  .  .  .
14 | .  D  .  D  <- (2, 3) aligned; 2*3 = 6 steps since previous
15 | D  .  D  .  <- (5, 2) aligned; 5*2 = 10 steps since previous
16 | .  D  .  .
17 | .  .  .  .
18 | .  D  D  .
19 | .  .  .  .
20 | D  D  .  .  <- (2, 3) aligned; 2*3 = 6 steps since previous
21 | .  .  D  D
22 | .  D  .  .
23 | .  .  .  .
24 | .  D  D  .
25 | D  .  .  .  <- (5, 2) aligned; 5*2 = 10 steps since previous
26 | .  D  .  .  <- (2, 3) aligned; 2*3 = 6 steps since previous
27 | .  .  D  .  <- (3, 7) aligned; 3*7 = 21 steps since previous
28 | .  D  .  D
29 | .  .  .  .
30 | D  D  D  .