2022 Day 11: Monkey in the Middle; part 2
It took me a long time to find how to reduce the worry number in a way
that kept its original properties.

I realised that the test divisors were primes. I first tried to
substitute the worry number by multiples of the primes up to and
including the highest test divisor, but this didn't give the correct
answer.

For a while I thought maybe I needed to take the operation into account.
I had almost given up and started skimming through the Reddit solutions
thread for this day to find hints. I didn't find any mention of the
operation, but many comments about primes and divisors which gave me
confidence back that I was on the right track. The thread also mentioned
the Chinese remainder theorem which I remember also reading about for
AoC 2020 day 13, but which I didn't (at least knowingly) use there.

This got me thinking about cycles (because I managed to get AoC 2020 day
13 right by examining cycles). I guessed that I probably needed to keep
the cycle offset (i.e. remainder) and that I could maybe safely remove
cycles if I just kept that intact. I realised that I needed to find a
common divisor and so I tried multiplying all of the test divisors
together and used this as the cycle. I subtracted the cycle from the
worry number in a loop until there wasn't enough left to subtract. This
made the tests pass. To my surprise, because I had been at this for so
long, trial & erroring my way forward, and felt I hadn't fully grasped
why it worked. At this point I noticed I had just implemented modulo and
switched out my loop for that.

Finally AoC 2022 day 11 is done. And it's almost noon.